[refactor] 메인, 로그인, 회원가입 파일 위치 수정
