+++ src/components/(main)/NewsData.tsx
