+++ src/components/(main)/RankingSection.tsx
